Move file mode audit to RuboCop.
