+++ Library/Homebrew/test/.rubocop_todo.yml
@@ -45,6 +45,7 @@ RSpec/FilePath:
     - 'rubocops/components_redundancy_spec.rb'
     - 'rubocops/conflicts_spec.rb'
     - 'rubocops/dependency_order_spec.rb'
+    - 'rubocops/files_spec.rb'
     - 'rubocops/homepage_spec.rb'
     - 'rubocops/options_spec.rb'
     - 'rubocops/patches_spec.rb'
